fine-tune-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fine-tune-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server"
authors = [
    { name = "rUv", email = "ruv@ruv.net" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["mcp>=1.0", "httpx>=0.24.0", "python-dotenv>=1.0.0"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
    "ruff>=0.2.0"
]

[tool.setuptools]
packages = ["fine_tune_mcp"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I"]
